requirements-exported.txt
annotated-types==0.6.0 ; python_version >= "3.12" and python_version < "3.13"
anyio==4.3.0 ; python_version >= "3.12" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.12" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "3.13" and platform_system == "Windows"
fastapi==0.110.0 ; python_version >= "3.12" and python_version < "3.13"
filelock==3.13.3 ; python_version >= "3.12" and python_version < "3.13"
fsspec==2024.3.1 ; python_version >= "3.12" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.12" and python_version < "3.13"
idna==3.6 ; python_version >= "3.12" and python_version < "3.13"
jinja2==3.1.3 ; python_version >= "3.12" and python_version < "3.13"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "3.13"
mpmath==1.3.0 ; python_version >= "3.12" and python_version < "3.13"
networkx==3.2.1 ; python_version >= "3.12" and python_version < "3.13"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "3.13"
nvidia-cublas-cu12==12.1.3.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-cuda-cupti-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-cuda-nvrtc-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-cuda-runtime-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-cudnn-cu12==8.9.2.26 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-cufft-cu12==11.0.2.54 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-curand-cu12==10.3.2.106 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-cusolver-cu12==11.4.5.107 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-cusparse-cu12==12.1.0.106 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-nccl-cu12==2.19.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-nvjitlink-cu12==12.4.99 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
nvidia-nvtx-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.13"
pillow==10.2.0 ; python_version >= "3.12" and python_version < "3.13"
pydantic-core==2.16.3 ; python_version >= "3.12" and python_version < "3.13"
pydantic==2.6.4 ; python_version >= "3.12" and python_version < "3.13"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "3.13"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "3.13"
starlette==0.36.3 ; python_version >= "3.12" and python_version < "3.13"
sympy==1.12 ; python_version >= "3.12" and python_version < "3.13"
torch==2.2.2 ; python_version >= "3.12" and python_version < "3.13"
torchvision==0.17.2 ; python_version >= "3.12" and python_version < "3.13"
typing-extensions==4.10.0 ; python_version >= "3.12" and python_version < "3.13"
uvicorn==0.29.0 ; python_version >= "3.12" and python_version < "3.13"
